requirements.txt
Flask==1.1.2
python-dotenv==0.14.0
faker==4.1.2
